requirements.txt
fastapi
uvicorn==0.34.0
PyJWT==2.10.1
dotenv==0.9.9
passlib==1.7.4
python-multipart==0.0.20
streamlit==1.43.2